+++ lib/jekyll/page_without_a_file.rb
@@ -1,0 +1,7 @@
+module Jekyll
+  class PageWithoutAFile < Page
+    def read_yaml(*)
+      @data ||= {}
+    end
+  end
+end
